✨ feat : ShowSchedule
